fix: algorithm to find normalized-hamming-distance
I was doing Cartesian product of the chunks of cipher buffer, which
isn't needed. We just want to find hamming distance of adjoining
chunks of key_size length. We can perhaps improve the algorithm by
checking [[1, 2], [2, 3], [3, 4]] instead of just [[1, 2], [3, 4]];
but I am getting good results with current algorithm so I am leaving
it at that.