fix: remove highlight.js based syntax highlighting
We are actually still using highlight.js CSS, but not its javascript
for syntax highlighting. We are kinda abusing org-mode to export code
blocks such that highlight.js' css can just style it. Some is lost in
this because the classes exported by org-mode aren't the same which
highlight.js expects. Once I've decided on a theme, I'll port it to
org-mode compatible and we'll have proper, very lightweight syntax
highlighting. Or we'll just use prism.js. But for now, this works.