wip: better tutorial-ish-ness of readme
challenge3 in readme.org and challenge3.rs is out of sync right now.
My focus is to get the <ob-cell> working well enough to support rich
exploratory programming which can also be published to web while
maintaining its riches. Right now, things from challenge3.rs need to
be brought back to readme.org (with org-detangle), but I am not
putting time in that right now. Not until <ob-cell> is to my
satisfaction.