fix: copy and modify old code from @observablehq/parser to parse module
instead of just splitting it at new-lines and calling it a collection of cells.
Although that was working without a problem. Not sure why this was needed.

I am trying to take the safest approach to keep my focus on building the
org-mode integration. This whole copied "Interpreter" is a massive hack on many
levels. I am keeping it as an unknown tech-debt; unknown because I don't know
what will be the payment for this.