refactor: use text-area for code blocks in HTML
It don't have syntax highlighting and all, but it is lightweight af. I tried
adding codemirror, but it is a whooping 500KB. I need to find a way to remove
acorn.js; which is around 110KB, while rest of my app is much lighter. May be
some other day.

I will perhaps add codemirror as an add-on via web-workers or something later.
It is a nice to have, but I don't want it to slow down the main app.